adding size_t type to sort interface.

+++ src/sort.hpp
@@ -1,1 +1,2 @@
-int* insertion_sort(int arr[], int len);
+#include <stddef.h>
+int* insertion_sort(int arr[], size_t len);
